requirements.txt
numpy
Django
djangorestframework
django-filter
django-cors-headers
django-bootstrap-ui
django-bootstrap3
markdown
pandas
pyfits
astropy
requests
astroquery
scipy
pyraf
pymongo
matplotlib
fitsio
cython
pyyaml
future
starlink-pyast
pymongo
pyvo